+++ src/NuGet.Clients/NuGet.CommandLine/NuGet.CommandLine.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <Import Project="..\..\..\build\Common.props" Condition="Exists('..\..\..\Build\Common.props')" />
@@ -81,6 +81,7 @@
     <Compile Include="ICommandManager.cs" />
     <Compile Include="MSBuildCachedRequestProvider.cs" />
     <Compile Include="MSBuildTasks\ProjectReferencesTask.cs" />
+    <Compile Include="MsbuildToolSet.cs" />
     <Compile Include="MsBuildUtility.cs" />
     <Compile Include="NuGetCommand.Designer.cs">
       <AutoGen>True</AutoGen>
